medi-fit/requirements.txt
aiohttp==3.10.5
python-dotenv==1.1.1
httpx==0.27.2
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.2
loguru==0.7.2
google-adk==0.2.0
twilio==9.8.3

cerebras_cloud_sdk==1.50.1
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6 
deepgram-sdk==2.12.0
numpy==1.26.0
google-cloud-texttospeech==2.14.1
requests==2.31.0

streamlit>=1.28.0
google-generativeai>=0.3.0
Pillow>=10.0.0
opencv-python>=4.8.0
pandas>=2.0.0
plotly>=5.17.0
pytesseract>=0.3.10
langdetect>=1.0.9
googletrans==4.0.0rc1
schedule>=1.2.0


